Move common code in lib for examples
